Remove Python dependency from zephyr-crosstool packages
This commit removes the Python dependency from the zephyr-crosstool
packages since the primary GDB executable now comes without Python
scripting support and therefore does not depend on the libpython
dynamic library.

The GDB Python scripting feature is now optional and may be utilised by
launching the *-gdb-py executable instead in which case an applicable
libpython must be available in the system.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>